Added retaining_default_executor
Revisiting issue #197

Having the queued tasks retain the executor is a completely seperate
problem from simply having the executor retained by those that are
referencing it. This fix prevents the crash outlined in issue #197 but doesn't
change the tasks that are dispatched.

If a user requires that the executor lifetime be well defined then they can
use `retaining_default_executor` otherwise `default_executor` will suffice.

- The `task_system` is now reference counted
- The `STLAB_TASK_SYSTEM_LIBDISPATCH` now uses the
  `task_system` model
- The new `retaining_default_executor` uses the
  same `task_system` as `default_executor`
- Only `STLAB_TASK_SYSTEM_PORTABLE`, `STLAB_TASK_SYSTEM_WINDOWS`,
  and `STLAB_TASK_SYSTEM_LIBDISPATCH` need to be reference counted
  so added and alias to the existing `default_executor` for the
  remaining systems