Always use Boost.Variant and Boost.Optional
Because we need to support C++14 compilers, Boost.Variant and Boost.Optional are
the only implementations of this functionality that support all our compilers.
This change substantially reduces complexity and ongoing maintenance costs.

Once we move to support only C++17, our intent is to switch to std::variant and
std::optional.