Use vanilla CMake for Windows CI
Using vcpkg to bring in Boost. Alternatives considered were using
chocolatey and building Boost from source. The former meadured more than
10 minutes to download and the latter added complications that were
unique to Boost.

Used Ninja as a generator. This required utilization of vcvarsall.bat,
but is (a) the fastest built tool, and (b) is used on the other
platforms as well.