Remove unused 'CMAKE_THREAD_PREFER_PTHREAD' variable. (#437)
The FindThreads CMake module is not impacted by setting this variable as the
original author likely intended. They were probably trying to set
THREADS_PREFER_PTHREAD_FLAG instead. However, due to a lack of rational for
attempting to set this variable, the line is removed here to restore the
default.