WASM CI (#445)
This PR adds CI support for Webassembly. It resolves #404, #405, and part of #406.

Note: 406 still requires unit tests for the WASM `main_executor` to be considered complete. 

This introduces a CMake toolchain file, `cmake/Platform/Emscripten-STLab.cmake` which extends the Emscripten SDK's own toolchain, and enables some experimental features required by STLab tests to compile and run successfully. Those features are threading support and exception handling. 

Following our existing convention, the tests may be executed simply with `ctest` from the build directory. A node runner is used.

Note that Node 16.16.0 or newer is required for sufficient exception handling support. The new CMake toolchain will report an error if this requirement is not satisfied. To use an appropriate version, we must edit a configuration file `.emscripten` in the emsdk directory to point to the system installation of node (16.16.0 is installed for GH Actions).