Move *future.hpp files to their new hyde location
This includes `stlab/concurrency/future.hpp` and `stlab/concurrency/ready_future.hpp`

The files were moved with the script `renames.sh`, added to this commit. The script will be modified in later commits (for other migrations) and removed from the subsequent PR. This was done to track file identity across versions of Hyde, which produces different outputs in newer versions.