Update documentation to new Hyde output format (#484)
This change is the result of running `hyde --hyde-update` on each header file in `stlab`. 

The only interesting code change is to `generate_docs.sh`, which has been renamed to validate_docs.sh Updates can now be done with a `-u` flag.

A good effort was made to preserve as much of the existing documentation history as possible. The protocol was as follows:

1. Run `hyde --hyde-update` on a single header to get a sense for the change.
2. If the generated documentation referred to components not yet documented, skip to step 8.
3. Identify the file moves, and update the script `renames.sh` used exclusively for this work. See commit [422f9fcf](https://github.com/stlab/libraries/commit/422f9fcf6d5e6b1b84ff1051338af4b344ce1a80) as an example.
4. Revert the changes with `git checkout -- docs && git clean -fxd docs`
5. Run `renames.sh` to move the old documentation to the locations they _would_ have been at had they been generated by a modern version of Hyde.
6. Commit this change
7. Rerun `hyde --hyde-update`
8. Manually fix up YAML changes as needed, often adding fields like `owner`. 
9. Commit this change.

This PR does not attempt to fill in `__MISSING__` fields for components not yet documented; this PR only rearranges the existing documentation to be consistent with what Hyde's latest version expects.