Revisit Reduce
I reimplemented the reduce mechanism as a (manually written) coroutine and added reduce as a feature to async.

The executor is now set correctly on the reduced future so continuations will execute where the wrapping future would have executed.

Reduce is now an independent algorithm but not exposed as the intention is to always auto-reduce. The