Removing option for library to reduce matrix
Since we have moved to CMake I'm fine with moving away from a header only library so I removed the header-only option.

I declared pre_exit and at_pre_exit as extern "C" to make them simpler to export from a DLL in an export list.

I will open an issue to do a similar thing for the default executor, that is a bit more difficult pull because of late type erasure.