No longer a header-only library. Added support for src files to make it simpler to integrate with dynamic libraries with hidden symbols.
Added .cpp files to hold thread pool and pre-exit support code to avoid duplicates when built as/in a DLL.

Declared pre_exit and at_pre_exit as extern "C" to make them simpler to export from a DLL in an export list.

Adding an issue to make it simpler to export the default executor as well.

---------

Co-authored-by: Guillaume BOTTESI <bottesi@adobe.com>
Co-authored-by: Sean Parent <sean.parent@stlab.cc>