libdispatch system_timer fix (#574)
* libdispatch system_timer fix

Once a timer is registered under libdispatch, we don't have provided way to revoke it. This adds the following:

Using the system timer and not invoking pre_exit() will result in an assertion.
pre_exit() will ensure any timer tasks that have not started will not execute after the call to pre_exit().
If the timer task has started to execute, it will be waited on during pre_exit().

The prior behavior was to await the completion of all scheduled tasks during pre_exit() - but that meant a task scheduled to execute tomorrow would appear to hang on quit.

* Adding system_timer_test for a timer scheduled after quit.